requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.16.0
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
lxml==5.1.0

